Reorganize utils with separate module for local paths
